Add files via upload
This c# code demonstrates the advantage of delegates and events. Using delegates and events we can create loosely coupled systems where objects of different classes communicate through events. So when we want create a new logic we don't really need to modify the existing logic in this case the 'VideoEncoder' logic.